requirements.txt
firebase_admin==6.5.0
geopy==2.4.1
python-telegram-bot==13.3
Requests==2.31.0
